requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
boto3==1.34.0
celery==5.3.4
redis==5.0.1
pydantic-settings==2.1.0
python-multipart==0.0.6